Refactor new geometry class file name
