+++ mbirjax/__init__.py
@@ -12,4 +12,4 @@ from .viewer import *
 from . import preprocess
 from .translation_model import *
 from .hsnt import *
-from .multiaxis_parallel_beam import *
+from .multiaxis_parallel import *
